requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
amberlite=16.0=pypi_0
ambertools=17.0=pypi_0
apbs=1.5=h14c3975_3
async-timeout=5.0.1=pypi_0
attrs=24.2.0=pypi_0
biopython=1.83=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_9
brotli-bin=1.0.9=h5eee18b_9
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.1.1=pypi_0
curl=8.9.1=hdbd6064_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
expat=2.6.3=h6a678d5_0
filelock=3.16.1=pypi_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.55.0=pypi_0
freemol=1.158=py_2
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
glew=2.0.0=0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.7.0=py38h737f45e_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.7=py38h06a4308_0
importlib-resources=6.4.5=pypi_0
importlib_resources=6.4.0=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libblas=3.9.0=1_h6e990d7_netlib
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcblas=3.9.0=3_h893e4fe_netlib
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.9.1=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.78.4=hdc74915_0
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1
libholoplaycore=0.1.0_rc4=1
libiconv=1.16=h5eee18b_3
liblapack=3.9.0=3_h893e4fe_netlib
libllvm14=14.0.6=hecde1de_4
libnetcdf=4.8.1=h8322cc2_2
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.18=hf726d26_0
libpng=1.6.39=h5eee18b_0
libpq=12.20=hdbd6064_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libzip=1.8.0=h6ac8c49_1
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-base=3.7.2=py38h1128e8f_0
mengine=1=h14c3975_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mmpbsa-py=16.0=pypi_0
mpeg_encode=1=h14c3975_1
mpmath=1.3.0=pypi_0
mtz2ccp4_px=1.0=h9ac9557_3
multidict=6.1.0=pypi_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openbabel=3.1.1=py38hd2c4bc0_2
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=py38h06a4308_0
packmol-memgen=1.0.5rc0=pypi_0
pandas=2.0.3=pypi_0
parmed=3.2.0=pypi_0
pcre2=10.42=hebb0a14_1
pdb2pqr=2.1.2+pymol=py_0
pdb4amber=1.7.dev0=pypi_0
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
ply=3.11=py38_0
pmw=2.0.1+3=py_3
propcache=0.2.0=pypi_0
psutil=6.1.0=pypi_0
pymol=2.4.1=py38h4463551_0
pyparsing=3.1.4=pypi_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0post0=py38h06a4308_2
python_abi=3.8=2_cp38
pytraj=2.0.5=pypi_0
pytz=2024.2=pypi_0
qt-main=5.15.2=h7358343_9
rdkit=2024.3.5=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
rigimol=1.3=2
sander=16.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py38h06a4308_0
torch=2.4.1=pypi_0
torch-geometric=2.6.1=pypi_0
tqdm=4.67.1=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.3=py38h06a4308_0
wheel=0.44.0=py38h06a4308_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.4.6=h5eee18b_1
yarl=1.15.2=pypi_0
zipp=3.20.2=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
